requirements.txt
fastapi>=0.115.0,<0.136
uvicorn[standard]>=0.32.0,<0.33
pydantic>=2.10.0,<3
yt-dlp>=2025.1.1
opencv-python-headless>=4.10.0,<5
Pillow>=10.4.0,<12
torch>=2.6.0
torchvision>=0.21.0
transformers>=4.46.0
numpy>=1.26.4,<3
huggingface-hub>=0.26.0
librosa>=0.10.0
soundfile>=0.12.0
python-multipart>=0.0.9
